backend/requirements.txt
Flask>=2.3.0
firebase-admin>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
recipe-scrapers>=14.40.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
flask-cors>=4.0.0
pytest>=7.4.0
gunicorn>=20.1.0
redis>=4.3.0
flask-swagger-ui>=4.11.1
apispec>=6.0.0
marshmallow>=3.19.0
marshmallow>=3.19.0
flask-restful>=0.3.9
flask-migrate>=4.0.0
python-dateutil>=2.8.0
